Transpilation: modify algorithm, add variables and number literals
The algorithm now uses a series of arrays to keep track of nested levels, rather than strings. Each array contains all individual objects on each level; when the level is ended, the array is joined with commas and merged into the previous level as part of an array or method call.